requirements.txt

# Sky Shield GUI - Minimal Requirements
# Sadece temel çalışma için gerekli paketler

customtkinter>=5.2.0
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0
websocket-client==1.6.3
websockets>=11.0.0